data/requirements.txt
google-genai>=1.10.0
weaviate-client>=4.0
yt-dlp>=2024.11.4
opencv-python-headless>=4.9.0.80
beautifulsoup4>=4.12.0
requests>=2.32.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
